Ref: extracting createApiResponse() from helpers.
