+++ src/index.ts
@@ -2,7 +2,8 @@ export { createConfig, LoggerConfig, ConfigType } from './config-type';
 export { AbstractEndpoint, EndpointInput, EndpointOutput, EndpointResponse } from './endpoint';
 export { Method } from './method';
 export { EndpointsFactory, defaultEndpointsFactory } from './endpoints-factory';
-export { IOSchema, FlatObject, markOutput, createApiResponse } from './helpers';
+export { IOSchema, FlatObject, markOutput} from './helpers';
+export { createApiResponse } from './api-response';
 export { createLogger } from './logger';
 export { createMiddleware } from './middleware';
 export { createResultHandler, defaultResultHandler} from './result-handler';
