+++ src/result-handler.ts
@@ -1,9 +1,8 @@
 import {Request, Response} from 'express';
 import {Logger} from 'winston';
 import {z} from 'zod';
+import {ApiResponse, createApiResponse} from './api-response';
 import {
-  ApiResponse,
-  createApiResponse,
   getMessageFromError,
   getStatusCodeFromError,
   IOSchema,
